requirements.txt
numpy
pandas
setuptools
python-dotenv
mysql-connector-python
pymysql
scikit-learn
cryptography
# -e .